fix: set_auto_lock_minutes accepte f64 au lieu de u32
La migration parseInt → parseFloat côté prefs.svelte.ts permet au front
d'envoyer 0.05 min au backend, mais la command Tauri attendait encore
u32 : la désérialisation échouait silencieusement (catch en console.warn),
laissant le watchdog tokio sur None. Le timer JS couvre déjà le cas en
E2E, mais en prod le watchdog backup ne tournait plus pour les fenêtres
courtes (et ne tournerait pas non plus si un jour on laisse l'utilisateur
saisir des demi-minutes).

Passage à Option<f64> dans AppState + Duration::from_secs_f64 dans le
watchdog. Le filtrage `minutes > 0.0 && minutes.is_finite()` remplace
l'ancien `minutes == 0` et protège aussi contre les NaN. Aucune régression
E2E (auto-lock passe toujours en ~7 s).

+++ src-tauri/src/state.rs
@@ -18,9 +18,11 @@ pub struct AppState {
     /// disabled JS timer can't keep the vault unlocked indefinitely.
     pub last_activity: Mutex<Instant>,
     /// `Some(n)` enables the auto-lock watchdog with an `n`-minute idle
-    /// window. `None` disables it. The frontend keeps this in sync via the
-    /// `set_auto_lock_minutes` command.
-    pub auto_lock_minutes: Mutex<Option<u32>>,
+    /// window. `None` disables it. Stored as `f64` to accommodate
+    /// sub-minute values written by the E2E suite via localStorage; the
+    /// production UI only ever writes positive integers. The frontend
+    /// keeps this in sync via the `set_auto_lock_minutes` command.
+    pub auto_lock_minutes: Mutex<Option<f64>>,
 }
 
 impl Default for AppState {
